+++ lib/glimmer/web/component/component_style_container.rb
@@ -7,7 +7,7 @@ class ComponentStyleContainer
   
   markup {
     component_style_container_block = component.class.instance_variable_get("@style_block")
-    component_style_container_class = "#{component.class.keyword.gsub('_' , '-')} #{component.markup_root.element_id}-style"
+    component_style_container_class = "#{component.class.component_element_class}-style"
     style(class: component_style_container_class, &component_style_container_block)
   }
 end