+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Optimize performance of Todo MVC by not adding an edit input field to every todo, yet adding it only upon editing a todo.
 - Upgrade to Glimmer DSL for CSS 1.5.0
 - Support `style {}` block in `Glimmer::Web::Component` that would automatically add style in one place for all components, without repeating style for repeating components
+- Remove support for including multiple `before_render` and `after_render` blocks in a component as it is not needed and can be confusing
 
 ## 0.3.2