+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Support `style {}` block in `Glimmer::Web::Component` that would automatically add style in one place for all components, without repeating style for repeating components
 - Remove support for including multiple `before_render` and `after_render` blocks in a component as it is not needed and can be confusing
 - Hello, Style! Sample: `require 'glimmer-dsl-web/samples/hello/hello_style'`
+- Fix issue with `ElementProxy#add_css_class` and `ElementProxy#remove_css_class` crashing if called before rendering an element
 
 ## 0.3.2